+++ src/main/java/theking530/staticpower/data/tiers/StaticPowerTierStatic.java
@@ -182,12 +182,12 @@ public class StaticPowerTierStatic extends StaticPowerTier {
 	}
 
 	@Override
-	public double getHeatCapacityUpgrade() {
+	public float getHeatCapacityUpgrade() {
 		return 1.0f;
 	}
 
 	@Override
-	public double getHeatConductivityUpgrade() {
+	public float getHeatConductivityUpgrade() {
 		return 1.0f;
 	}
